Share t_safe functions
